+++ packages/cortex-react/docs/select-multi-grouped.stories.tsx
@@ -1,6 +1,5 @@
 import { StoryFn } from '@storybook/react';
-import { checkbox } from '@tecsinapse/cortex-core';
-import React, { ChangeEvent, useCallback, useState } from 'react';
+import React, { useState } from 'react';
 import { Input, Select } from '../src';
 
 export default {
@@ -49,9 +48,8 @@ const map = new Map([
 ]);
 
 const Template: StoryFn = args => {
-  const [value, setValue] = useState<{ key: string; label: string }[]>();
+  const [value, setValue] = useState<{ key: string; label: string }[]>([]);
   const [options, setOptions] = useState(map);
-  const flattenMap = Array.from(map.values()).flatMap(value => value)
 
   const handleSearch = React.useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     const searchArg = event.target.value;
@@ -66,15 +64,6 @@ const Template: StoryFn = args => {
     setOptions(newMap)
   }, [options]);
 
-  const checkAll = useCallback((e: ChangeEvent<HTMLInputElement>) => {
-    {
-        if(!e.target?.checked)
-            setValue([])
-        else 
-            setValue(flattenMap)
-    }
-  }, [])
-
   return (
     <div className={'w-[350px] h-[250px]'}>
       <Select.Root 
@@ -84,24 +73,20 @@ const Template: StoryFn = args => {
       >
         <Select.Trigger label={args.label} />
         <Select.Popover>
-          <div className='flex flex-row items-center gap-x-deca px-deca'>
-                <input
-                    type='checkbox' 
-                    className={checkbox()}
-                    onChange={checkAll}
-                    checked={flattenMap.length === value?.length}
-                />
-                <Input.Search 
-                    className={'flex-1'} 
-                    onChange={handleSearch} 
-                    placeholder={args.placeholderSearchInput} 
-                />
-          </div>
+          <Input.Search 
+            variants={{
+              className:'flex-1 mx-deca mt-centi' 
+            }}
+            onChange={handleSearch} 
+            placeholder={args.placeholderSearchInput} 
+          />
           <Select.MultiGroupedOptions
             groupedLabelExtractor={labelGroup => labelGroup} 
             options={options} 
             onSelect={(option) => setValue(option)}
-          />
+          >
+            <Select.MultiCheckAllOptions />
+          </Select.MultiGroupedOptions>
         </Select.Popover>
       </Select.Root>
     </div>