+++ packages/cortex-react/src/components/Select/MultiCheckAllOptions.tsx
@@ -1,0 +1,55 @@
+import { checkbox, option } from '@tecsinapse/cortex-core';
+import React, { useCallback, useContext, useMemo } from 'react';
+import { SelectContext, SelectMultiOptionsContext } from './context';
+
+export interface SelectMultiCheckAllOptionsProps {
+  checkAllLabel?: string;
+}
+
+export const SelectMultiCheckAllOptions = <T,>({
+  checkAllLabel = 'Select all',
+}: SelectMultiCheckAllOptionsProps) => {
+  const { keyExtractor, value: currentValue = [] } = useContext(SelectContext);
+  const { options, onSelect } = useContext(SelectMultiOptionsContext);
+
+  const isChecked = useMemo(
+    () =>
+      (options ?? []).filter(it =>
+        currentValue.find(value => keyExtractor(value) === keyExtractor(it))
+      ).length == (options ?? []).length,
+    [keyExtractor, currentValue, options]
+  );
+
+  const checkAll = useCallback(() => {
+    {
+      let updateSelected: T[] = [...currentValue];
+      if (!isChecked) {
+        updateSelected = [
+          ...new Set<T>([...updateSelected, ...(options ?? [])]),
+        ];
+      } else {
+        const optionKeys = new Set(
+          options?.map(option => keyExtractor(option))
+        );
+        updateSelected = updateSelected.filter(
+          item => !optionKeys.has(keyExtractor(item))
+        );
+      }
+      onSelect(updateSelected);
+    }
+  }, [options, onSelect]);
+
+  return options?.length ? (
+    <div className={option()} onClick={checkAll}>
+      <input
+        type="checkbox"
+        className={checkbox()}
+        onChange={checkAll}
+        checked={isChecked}
+      />
+      <span>{checkAllLabel}</span>
+    </div>
+  ) : (
+    <></>
+  );
+};