+++ packages/cortex-react/src/components/Select/MultiOption.tsx
@@ -5,11 +5,13 @@ import { SelectContext } from './context';
 export interface SelectMultiOptionProps<T> {
   option: T;
   onSelectOption: (option: T) => void;
+  grouped?: boolean;
 }
 
 export const SelectMultiOption = <T,>({
   onSelectOption,
   option,
+  grouped = false,
 }: SelectMultiOptionProps<T>) => {
   const { keyExtractor, labelExtractor, value } = useContext(SelectContext);
   const inputRef = useRef<HTMLInputElement>(null);
@@ -28,11 +30,16 @@ export const SelectMultiOption = <T,>({
   }, [onSelectOption, inputRef]);
 
   return (
-    <li onClick={onClickOption} className={styleOption()} role={'option'}>
+    <li
+      onClick={onClickOption}
+      className={styleOption({ grouped })}
+      role={'option'}
+    >
       <input
         type="checkbox"
         checked={isChecked}
         className={checkbox({ className: 'pointer-events-none' })}
+        onChange={() => false}
         ref={inputRef}
       />
       {labelExtractor(option)}